perf(app): code split degen browser route (#516)
